+++ leetcode/2021/february/practice/java/1480-RunningSumof1dArray.java
@@ -1,0 +1,15 @@
+import java.util.Arrays;
+
+class Solution {
+    public static int[] runningSum(int[] nums) { 
+        for(int i=1;i<nums.length;i++){
+            nums[i]=nums[i-1]+nums[i];
+        }
+        return nums;
+    }
+
+    public static void main(String[] args) {
+        int[] arr = {1,2,3,4};
+        System.out.println(Arrays.toString(runningSum(arr)));;
+    }
+}